♻️ feat: handling error
