+++ src/lib.rs
@@ -1,3 +1,5 @@
 mod constants;
 pub mod finary;
 pub mod sign_in;
+pub mod error;
+pub mod structs;
